Fake-backend-vue/fakebackend/requirements.txt
Django
django-cors-headers
django-extensions
djangorestframework